requirements.txt
# pytest
asyncio
aiohttp
cachetools
python-dotenv
fastapi
uvicorn[standard]




#test
#manejo de errores
#comentarios del codigo
#organizar docker
# cachetools
